requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
sentence-transformers==2.6.0
faiss-cpu==1.7.4
numpy==1.26.4
python-multipart==0.0.7
